Extract retry constant and method into HadoopTestUtils
