+++ testing/trino-product-tests/src/main/java/io/trino/tests/product/hive/HiveProductTest.java
@@ -13,45 +13,13 @@
  */
 package io.trino.tests.product.hive;
 
-import com.google.common.base.Throwables;
-import io.airlift.log.Logger;
 import io.trino.tempto.ProductTest;
-import io.trino.tempto.query.QueryResult;
-import net.jodah.failsafe.RetryPolicy;
-import org.intellij.lang.annotations.Language;
 
 import javax.inject.Inject;
-
-import java.time.temporal.ChronoUnit;
-import java.util.regex.Pattern;
 
 public class HiveProductTest
         extends ProductTest
 {
-    private static final Logger log = Logger.get(HiveProductTest.class);
-
-    static final String ERROR_COMMITTING_WRITE_TO_HIVE_ISSUE = "https://github.com/trinodb/trino/issues/4936";
-    @Language("RegExp")
-    static final String ERROR_COMMITTING_WRITE_TO_HIVE_MATCH =
-            // "Error committing write to Hive" is present depending on when the exception is thrown.
-            // It may be absent when the underlying problem manifest earlier (e.g. during RecordFileWriter.appendRow vs RecordFileWriter.commit).
-
-            // "could only be written to 0 of the 1 minReplication" is the error wording used by e.g. HDP 3
-            "(could only be replicated to 0 nodes instead of minReplication|could only be written to 0 of the 1 minReplication)";
-
-    public static final RetryPolicy<QueryResult> ERROR_COMMITTING_WRITE_TO_HIVE_RETRY_POLICY = new RetryPolicy<QueryResult>()
-            .handleIf(HiveProductTest::isErrorCommittingToHive)
-            .withBackoff(1, 10, ChronoUnit.SECONDS)
-            .withMaxRetries(30)
-            .onRetry(event -> log.warn(event.getLastFailure(), "Query failed on attempt %d, will retry.", event.getAttemptCount()));
-
-    private static boolean isErrorCommittingToHive(Throwable throwable)
-    {
-        return Pattern.compile(ERROR_COMMITTING_WRITE_TO_HIVE_MATCH)
-                .matcher(Throwables.getStackTraceAsString(throwable))
-                .find();
-    }
-
     @Inject
     private HiveVersionProvider hiveVersionProvider;
